+++ hslua-packaging/test/HsLua/Packaging/RenderingTests.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeApplications #-}
+{-# OPTIONS_GHC -Wno-warnings-deprecations #-}
 {-|
 Module      : HsLua.Packaging.RenderingTests
 Copyright   : © 2020-2022 Albert Krewinkel